use what Ethan already made as a basis to semisupervisedmixin
